+++ docker/.env.ai.example
@@ -1,26 +1,36 @@
 ## LLM
-LLM_PROVIDER= # openai, azure_openai, ollama
+LLM_PROVIDER= # openai_llm, azure_openai_llm, ollama_llm
+GENERATION_MODEL=
 
-# openai or openai-api-compatible llm
-OPENAI_API_KEY=
-OPENAI_API_BASE=
+# openai or openai-api-compatible
+LLM_OPENAI_API_KEY=
+LLM_OPENAI_API_BASE=
 
 # azure_openai
-AZURE_CHAT_BASE=
-AZURE_CHAT_KEY=
-AZURE_CHAT_VERSION=
-
-AZURE_EMBED_BASE=
-AZURE_EMBED_KEY=
-AZURE_EMBED_VERSION=
+LLM_AZURE_OPENAI_API_KEY=
+LLM_AZURE_OPENAI_API_BASE=
+LLM_AZURE_OPENAI_VERSION=
 
 # ollama
-OLLAMA_URL=http://host.docker.internal:11434
+LLM_OLLAMA_URL=http://localhost:11434
 
-GENERATION_MODEL=
-# supported embedding models providers by qdrant: https://qdrant.tech/documentation/embeddings/
+
+## EMBEDDER
+EMBEDDER_PROVIDER= # openai_embedder, azure_openai_embedder, ollama_embedder
 EMBEDDING_MODEL=
 EMBEDDING_MODEL_DIMENSION=
+
+# openai or openai-api-compatible
+EMBEDDER_OPENAI_API_KEY=
+EMBEDDER_OPENAI_API_BASE=
+
+# azure_openai
+EMBEDDER_AZURE_OPENAI_API_KEY=
+EMBEDDER_AZURE_OPENAI_API_BASE=
+EMBEDDER_AZURE_OPENAI_VERSION=
+
+# ollama
+EMBEDDER_OLLAMA_URL=http://localhost:11434
 
 
 ## DOCUMENT_STORE
