+++ docker/docker-compose-dev.yaml
@@ -44,10 +44,10 @@ services:
     environment:
       WREN_AI_SERVICE_PORT: ${WREN_AI_SERVICE_PORT}
       WREN_UI_ENDPOINT: ${WREN_UI_ENDPOINT}
-      GENERATION_MODEL: ${GENERATION_MODEL}
-      OPENAI_API_KEY: ${OPENAI_API_KEY}
-      AZURE_CHAT_KEY: ${AZURE_CHAT_KEY}
-      AZURE_EMBED_KEY: ${AZURE_EMBED_KEY}
+      LLM_OPENAI_API_KEY: ${LLM_OPENAI_API_KEY}
+      EMBEDDER_OPENAI_API_KEY: ${EMBEDDER_OPENAI_API_KEY}
+      LLM_AZURE_OPENAI_API_KEY: ${LLM_AZURE_OPENAI_API_KEY}
+      EMBEDDER_AZURE_OPENAI_API_KEY: ${EMBEDDER_AZURE_OPENAI_API_KEY}
       ENABLE_TIMER: ${AI_SERVICE_ENABLE_TIMER}
       LOGGING_LEVEL: ${AI_SERVICE_LOGGING_LEVEL}
       # sometimes the console won't show print messages,
