+++ wren-ai-service/.env.dev.example
@@ -5,39 +5,52 @@ WREN_ENGINE_ENDPOINT=http://localhost:8080
 WREN_UI_ENDPOINT=http://localhost:3000
 
 ## LLM
-LLM_PROVIDER=openai # openai, azure_openai, ollama
+LLM_PROVIDER=openai_llm # openai_llm, azure_openai_llm, ollama_llm
+GENERATION_MODEL=gpt-3.5-turbo
 
-# openai or openai-api-compatible llm
-OPENAI_API_KEY=sk-1234567890
-OPENAI_API_BASE=https://api.openai.com/v1
+# openai or openai-api-compatible
+LLM_OPENAI_API_KEY=sk-1234567890
+LLM_OPENAI_API_BASE=https://api.openai.com/v1
 
 # azure_openai
-AZURE_CHAT_BASE=
-AZURE_CHAT_KEY=
-AZURE_CHAT_VERSION=
-
-AZURE_EMBED_BASE=
-AZURE_EMBED_KEY=
-AZURE_EMBED_VERSION=
+LLM_AZURE_OPENAI_API_KEY=
+LLM_AZURE_OPENAI_API_BASE=
+LLM_AZURE_OPENAI_VERSION=
 
 # ollama
-OLLAMA_URL=http://localhost:11434
+LLM_OLLAMA_URL=http://localhost:11434
 
-GENERATION_MODEL=gpt-3.5-turbo
+
+## EMBEDDER
+EMBEDDER_PROVIDER=openai_embedder # openai_embedder, azure_openai_embedder, ollama_embedder
 EMBEDDING_MODEL=text-embedding-3-large
 EMBEDDING_MODEL_DIMENSION=3072
+
+# openai or openai-api-compatible
+EMBEDDER_OPENAI_API_KEY=sk-1234567890
+EMBEDDER_OPENAI_API_BASE=https://api.openai.com/v1
+
+# azure_openai
+EMBEDDER_AZURE_OPENAI_API_KEY=
+EMBEDDER_AZURE_OPENAI_API_BASE=
+EMBEDDER_AZURE_OPENAI_VERSION=
+
+# ollama
+EMBEDDER_OLLAMA_URL=http://localhost:11434
+
 
 ## DOCUMENT_STORE
 DOCUMENT_STORE_PROVIDER=qdrant
 
 QDRANT_HOST=http://localhost:6333
 
-ENGINE=wren-ui
+## ENGINE
+ENGINE=wren_ui
 
-## when using wren-ui as the engine
+## when using wren_ui as the engine
 WREN_UI_ENDPOINT=http://localhost:3000
 
-## when using ibis as the engine
+## when using wren_ibis as the engine
 WREN_IBIS_ENDPOINT=http://localhost:8000
 WREN_IBIS_SOURCE=bigquery
 WREN_IBIS_MANIFEST= # this is a base64 encoded string of the MDL
