+++ wren-ai-service/.env.example
@@ -1,3 +1,2 @@
-# in the future, we might delete this file since
-# we only need to use .env.dev or .env.prod
+# if not specified, the default ENV is prod
 ENV=dev
