--- wren-ai-service/.env.prod.example
@@ -1,42 +1,0 @@
-# docker related
-# This will determine the prefix of container name.
-# If the service name in docker-compose.yaml is ai-service,
-# then the container name will be wren-ai-service-1.
-# If COMPOSE_PROJECT_NAME is not set, the default prefix name is docker.
-COMPOSE_PROJECT_NAME=wren
-PLATFORM=linux/amd64
-
-# fastapi related
-WREN_AI_SERVICE_PORT=5555
-
-# app related
-# LLM Provider name should be mapped to Haystack's supported LLM providers: https://docs.haystack.deepset.ai/v2.0/docs/generators
-LLM_PROVIDER=openai # azure_openai as well
-
-
-# Document Store Provider name should be mapped to Haystack's supported Document Store providers: see the haystack documentation's left sidebar
-DOCUMENT_STORE_PROVIDER=qdrant
-
-# llm provider specific env variables names must be in the format of [LLM_PROVIDER]_[ENV_VARIABLE_NAME] and must be in uppercase
-OPENAI_API_KEY=
-OPENAI_API_BASE=https://api.openai.com/v1
-EMBEDDING_MODEL=text-embedding-3-large
-EMBEDDING_MODEL_DIMENSION=3072
-GENERATION_MODEL=gpt-3.5-turbo # gpt-4o, gpt-4-turbo, gpt-3.5-turbo
-
-#Azure openai env
-AZURE_CHAT_BASE=
-AZURE_CHAT_KEY=
-AZURE_CHAT_VERSION=
-
-AZURE_EMBED_BASE=
-AZURE_EMBED_KEY=
-AZURE_EMBED_VERSION=
-
-# document store provider specific env variables names must be in the format of [DOCUMENT_STORE_PROVIDER]_[ENV_VARIABLE_NAME] and must be in uppercase
-QDRANT_HOST=qdrant
-
-WREN_UI_ENDPOINT=
-
-ENABLE_TIMER=
-LOGGING_LEVEL=INFO
