+++ wren-ai-service/Makefile
@@ -13,17 +13,6 @@ dev-down:
 ## wren-ai-service related ##
 start:
 	poetry run python -m src.__main__
-
-build:
-	docker compose -f docker/docker-compose.yaml --env-file .env.prod build
-
-up:
-	make dev-up
-	docker compose -f docker/docker-compose.yaml --env-file .env.prod up -d
-
-down:
-	make dev-down
-	docker compose -f docker/docker-compose.yaml --env-file .env.prod down
 ## wren-ai-service related ##
