+++ wren-ai-service/src/providers/llm/openai.py
@@ -1,40 +1,36 @@
 import logging
 import os
-from typing import Any, Callable, Dict, List, Optional, Tuple, Union
+from typing import Any, Callable, Dict, List, Optional, Union
 
 import backoff
 import openai
-from haystack import Document, component
-from haystack.components.embedders import OpenAIDocumentEmbedder, OpenAITextEmbedder
+from haystack import component
 from haystack.components.generators import OpenAIGenerator
 from haystack.dataclasses import ChatMessage, StreamingChunk
 from haystack.utils import Secret
 from openai import AsyncOpenAI, OpenAI, Stream
 from openai.types.chat import ChatCompletion, ChatCompletionChunk
-from tqdm import tqdm
 
 from src.core.provider import LLMProvider
 from src.providers.loader import provider
 
 logger = logging.getLogger("wren-ai-service")
 
-OPENAI_API_BASE = "https://api.openai.com/v1"
-GENERATION_MODEL_NAME = "gpt-3.5-turbo"
+LLM_OPENAI_API_BASE = "https://api.openai.com/v1"
+GENERATION_MODEL = "gpt-3.5-turbo"
 GENERATION_MODEL_KWARGS = {
     "temperature": 0,
     "n": 1,
     "max_tokens": 4096,
     "response_format": {"type": "json_object"},
 }
-EMBEDDING_MODEL_NAME = "text-embedding-3-large"
-EMBEDDING_MODEL_DIMENSION = 3072
 
 
 @component
 class AsyncGenerator(OpenAIGenerator):
     def __init__(
         self,
-        api_key: Secret = Secret.from_env_var("OPENAI_API_KEY"),
+        api_key: Secret = Secret.from_env_var("LLM_OPENAI_API_KEY"),
         model: str = "gpt-3.5-turbo",
         streaming_callback: Optional[Callable[[StreamingChunk], None]] = None,
         api_base_url: Optional[str] = None,
@@ -116,174 +112,13 @@ class AsyncGenerator(OpenAIGenerator):
         }
 
 
-@component
-class AsyncTextEmbedder(OpenAITextEmbedder):
-    def __init__(
-        self,
-        api_key: Secret = Secret.from_env_var("OPENAI_API_KEY"),
-        model: str = "text-embedding-ada-002",
-        dimensions: Optional[int] = None,
-        api_base_url: Optional[str] = None,
-        organization: Optional[str] = None,
-        prefix: str = "",
-        suffix: str = "",
-    ):
-        super(AsyncTextEmbedder, self).__init__(
-            api_key,
-            model,
-            dimensions,
-            api_base_url,
-            organization,
-            prefix,
-            suffix,
-        )
-        self.client = AsyncOpenAI(
-            api_key=api_key.resolve_value(),
-            organization=organization,
-            base_url=api_base_url,
-        )
-
-    @component.output_types(embedding=List[float], meta=Dict[str, Any])
-    @backoff.on_exception(backoff.expo, openai.RateLimitError, max_time=60, max_tries=3)
-    async def run(self, text: str):
-        if not isinstance(text, str):
-            raise TypeError(
-                "OpenAITextEmbedder expects a string as an input."
-                "In case you want to embed a list of Documents, please use the OpenAIDocumentEmbedder."
-            )
-
-        logger.debug(f"Running Async OpenAI text embedder with text: {text}")
-
-        text_to_embed = self.prefix + text + self.suffix
-
-        # copied from OpenAI embedding_utils (https://github.com/openai/openai-python/blob/main/openai/embeddings_utils.py)
-        # replace newlines, which can negatively affect performance.
-        text_to_embed = text_to_embed.replace("\n", " ")
-
-        if self.dimensions is not None:
-            response = await self.client.embeddings.create(
-                model=self.model, dimensions=self.dimensions, input=text_to_embed
-            )
-        else:
-            response = await self.client.embeddings.create(
-                model=self.model, input=text_to_embed
-            )
-
-        meta = {"model": response.model, "usage": dict(response.usage)}
-
-        return {"embedding": response.data[0].embedding, "meta": meta}
-
-
-@component
-class AsyncDocumentEmbedder(OpenAIDocumentEmbedder):
-    def __init__(
-        self,
-        api_key: Secret = Secret.from_env_var("OPENAI_API_KEY"),
-        model: str = "text-embedding-ada-002",
-        dimensions: Optional[int] = None,
-        api_base_url: Optional[str] = None,
-        organization: Optional[str] = None,
-        prefix: str = "",
-        suffix: str = "",
-        batch_size: int = 32,
-        progress_bar: bool = True,
-        meta_fields_to_embed: Optional[List[str]] = None,
-        embedding_separator: str = "\n",
-    ):
-        super(AsyncDocumentEmbedder, self).__init__(
-            api_key,
-            model,
-            dimensions,
-            api_base_url,
-            organization,
-            prefix,
-            suffix,
-            batch_size,
-            progress_bar,
-            meta_fields_to_embed,
-            embedding_separator,
-        )
-        self.client = AsyncOpenAI(
-            api_key=api_key.resolve_value(),
-            organization=organization,
-            base_url=api_base_url,
-        )
-
-    async def _embed_batch(
-        self, texts_to_embed: List[str], batch_size: int
-    ) -> Tuple[List[List[float]], Dict[str, Any]]:
-        all_embeddings = []
-        meta: Dict[str, Any] = {}
-        for i in tqdm(
-            range(0, len(texts_to_embed), batch_size),
-            disable=not self.progress_bar,
-            desc="Calculating embeddings",
-        ):
-            batch = texts_to_embed[i : i + batch_size]
-            if self.dimensions is not None:
-                response = await self.client.embeddings.create(
-                    model=self.model, dimensions=self.dimensions, input=batch
-                )
-            else:
-                response = await self.client.embeddings.create(
-                    model=self.model, input=batch
-                )
-            embeddings = [el.embedding for el in response.data]
-            all_embeddings.extend(embeddings)
-
-            if "model" not in meta:
-                meta["model"] = response.model
-            if "usage" not in meta:
-                meta["usage"] = dict(response.usage)
-            else:
-                meta["usage"]["prompt_tokens"] += response.usage.prompt_tokens
-                meta["usage"]["total_tokens"] += response.usage.total_tokens
-
-        return all_embeddings, meta
-
-    @component.output_types(documents=List[Document], meta=Dict[str, Any])
-    @backoff.on_exception(backoff.expo, openai.RateLimitError, max_time=60, max_tries=3)
-    async def run(self, documents: List[Document]):
-        if (
-            not isinstance(documents, list)
-            or documents
-            and not isinstance(documents[0], Document)
-        ):
-            raise TypeError(
-                "OpenAIDocumentEmbedder expects a list of Documents as input."
-                "In case you want to embed a string, please use the OpenAITextEmbedder."
-            )
-
-        logger.debug(
-            f"Running Async OpenAI document embedder with documents: {documents}"
-        )
-
-        texts_to_embed = self._prepare_texts_to_embed(documents=documents)
-
-        embeddings, meta = await self._embed_batch(
-            texts_to_embed=texts_to_embed, batch_size=self.batch_size
-        )
-
-        for doc, emb in zip(documents, embeddings):
-            doc.embedding = emb
-
-        return {"documents": documents, "meta": meta}
-
-
-@provider("openai")
+@provider("openai_llm")
 class OpenAILLMProvider(LLMProvider):
     def __init__(
         self,
-        api_key: Secret = Secret.from_env_var("OPENAI_API_KEY"),
-        api_base: str = os.getenv("OPENAI_API_BASE") or OPENAI_API_BASE,
-        embedding_model: str = os.getenv("EMBEDDING_MODEL") or EMBEDDING_MODEL_NAME,
-        embedding_model_dim: int = (
-            int(os.getenv("EMBEDDING_MODEL_DIMENSION"))
-            if os.getenv("EMBEDDING_MODEL_DIMENSION")
-            else 0
-        )
-        or EMBEDDING_MODEL_DIMENSION,
-        generation_model: str = os.getenv("GENERATION_MODEL") or GENERATION_MODEL_NAME,
+        api_key: Secret = Secret.from_env_var("LLM_OPENAI_API_KEY"),
+        api_base: str = os.getenv("LLM_OPENAI_API_BASE") or LLM_OPENAI_API_BASE,
+        generation_model: str = os.getenv("GENERATION_MODEL") or GENERATION_MODEL,
     ):
         def _verify_api_key(api_key: str, api_base: str) -> None:
             """
@@ -293,17 +128,13 @@ class OpenAILLMProvider(LLMProvider):
 
         logger.info(f"Initializing OpenAILLM provider with API base: {api_base}")
         # TODO: currently only OpenAI api key can be verified
-        if api_base == OPENAI_API_BASE:
+        if api_base == LLM_OPENAI_API_BASE:
             _verify_api_key(api_key.resolve_value(), api_base)
-            logger.info(f"Using OpenAI Generation Model: {generation_model}")
+            logger.info(f"Using OpenAI LLM: {generation_model}")
         else:
-            logger.info(
-                f"Using OpenAI API-compatible Generation Model: {generation_model}"
-            )
+            logger.info(f"Using OpenAI API-compatible LLM: {generation_model}")
         self._api_key = api_key
         self._api_base = api_base
-        self._embedding_model = embedding_model
-        self._embedding_model_dim = embedding_model_dim
         self._generation_model = generation_model
 
     def get_generator(
@@ -317,20 +148,4 @@ class OpenAILLMProvider(LLMProvider):
             model=self._generation_model,
             system_prompt=system_prompt,
             generation_kwargs=model_kwargs,
-        )
-
-    def get_text_embedder(self):
-        return AsyncTextEmbedder(
-            api_key=self._api_key,
-            api_base_url=self._api_base,
-            model=self._embedding_model,
-            dimensions=self._embedding_model_dim,
-        )
-
-    def get_document_embedder(self):
-        return AsyncDocumentEmbedder(
-            api_key=self._api_key,
-            api_base_url=self._api_base,
-            model=self._embedding_model,
-            dimensions=self._embedding_model_dim,
         )
