+++ wren-ai-service/tests/pytest/pipelines/test_ask.py
@@ -4,7 +4,7 @@ import orjson
 import pytest
 
 from src.core.pipeline import async_validate
-from src.core.provider import DocumentStoreProvider, LLMProvider
+from src.core.provider import DocumentStoreProvider, EmbedderProvider
 from src.pipelines.ask.followup_generation import FollowUpGeneration
 from src.pipelines.ask.generation import Generation
 from src.pipelines.ask.retrieval import Retrieval
@@ -26,24 +26,31 @@ def mdl_str():
 
 @pytest.fixture
 def llm_provider():
-    llm_provider, _, _ = init_providers()
+    llm_provider, _, _, _ = init_providers()
 
     return llm_provider
 
 
 @pytest.fixture
+def embedder_provider():
+    _, embedder_provider, _, _ = init_providers()
+
+    return embedder_provider
+
+
+@pytest.fixture
 def document_store_provider():
-    _, document_store_provider, _ = init_providers()
+    _, _, document_store_provider, _ = init_providers()
 
     return document_store_provider
 
 
 def test_clear_documents(mdl_str: str):
-    llm_provider, document_store_provider, _ = init_providers()
+    _, embedder_provider, document_store_provider, _ = init_providers()
     store = document_store_provider.get_store()
 
     indexing_pipeline = Indexing(
-        llm_provider=llm_provider,
+        embedder_provider=embedder_provider,
         document_store_provider=document_store_provider,
     )
 
@@ -78,11 +85,11 @@ def test_clear_documents(mdl_str: str):
 
 def test_indexing_pipeline(
     mdl_str: str,
-    llm_provider: LLMProvider,
+    embedder_provider: EmbedderProvider,
     document_store_provider: DocumentStoreProvider,
 ):
     indexing_pipeline = Indexing(
-        llm_provider=llm_provider,
+        embedder_provider=embedder_provider,
         document_store_provider=document_store_provider,
     )
 
@@ -98,11 +105,11 @@ def test_indexing_pipeline(
 
 
 def test_retrieval_pipeline(
-    llm_provider: LLMProvider,
+    embedder_provider: EmbedderProvider,
     document_store_provider: DocumentStoreProvider,
 ):
     retrieval_pipeline = Retrieval(
-        llm_provider=llm_provider,
+        embedder_provider=embedder_provider,
         document_store_provider=document_store_provider,
     )
 
@@ -119,7 +126,7 @@ def test_retrieval_pipeline(
 
 
 def test_generation_pipeline():
-    llm_provider, _, engine = init_providers()
+    llm_provider, _, _, engine = init_providers()
     generation_pipeline = Generation(llm_provider=llm_provider, engine=engine)
     generation_result = async_validate(
         lambda: generation_pipeline.run(
@@ -146,7 +153,7 @@ def test_generation_pipeline():
 
 
 def test_followup_generation_pipeline():
-    llm_provider, _, engine = init_providers()
+    llm_provider, _, _, engine = init_providers()
     generation_pipeline = FollowUpGeneration(llm_provider=llm_provider, engine=engine)
     generation_result = async_validate(
         lambda: generation_pipeline.run(
@@ -172,7 +179,7 @@ def test_followup_generation_pipeline():
 
 
 def test_sql_correction_pipeline():
-    llm_provider, _, engine = init_providers()
+    llm_provider, _, _, engine = init_providers()
     sql_correction_pipeline = SQLCorrection(llm_provider=llm_provider, engine=engine)
 
     sql_correction_result = async_validate(
