+++ wren-ai-service/tests/pytest/services/test_ask_details.py
@@ -14,7 +14,7 @@ from src.web.v1.services.ask_details import (
 
 @pytest.fixture
 def ask_details_service():
-    llm_provider, _, engine = init_providers()
+    llm_provider, _, _, engine = init_providers()
     return AskDetailsService(
         {
             "generation": generation.Generation(
